Fix for issue #593
I also updated the docstring.

There was a typo in the regexp, where the : was in the wrong place I think. I
don't think this change will break anything, most places this regexp is used
only consider the whole match. In one place, there is a use of match-string 1 in
org-ref-next-key.